yanlış anlaşılma giderildi
